Bug 1619953 - P11 - Mochitest `test_presentation_dc_sender.html` disabled for e10s r=mccr8
The test was found to fail for various builds in try runs for Bug 1656033. Running manually on the latest mozilla central at this time, the test did not pass for Linux and Windows 64 bit debug. This should be more closely investigated at a later time.

Filed Bug 1656033 to re-enable.

Differential Revision: https://phabricator.services.mozilla.com/D85313

UltraBlame original commit: 82321c8cba4557efaeccddb06806bf0681288575