Bug 1648051 Task resource monitor r=mtabara,tomprince
I've left the monitor disabled for now, so that we can have a smaller pushes for enabling and disabling it if needed. It should allow more fine grained control.

We may also want to include extracting the monitor tool from a github version instead, and also removing the assumption and it being forked from the parent, so that it's instead given a process ID to treat as the parent it should watch.

Differential Revision: https://phabricator.services.mozilla.com/D84374

UltraBlame original commit: f8a90f470a9c910aacf73084f73f3963b382fc9c