Bug 1657231, part 2 - Clear sInitialized when we destroy the singleton url preloader. r=kmag
Some of the entry points into the URL preloader check sInitialized, and only
use it if it is true. However, if we're late in shutdown and we've already
cleared the singleton for the preloader then our failure to clear the
sInitialized flag means we'll recreate the preloader, which is going
to fail in various ways.

I think this won't help in the case of bug 1656973, because that seems
to go through AutoBeginReading, which unconditionally creates a new
singleton if one does not exist.

Differential Revision: https://phabricator.services.mozilla.com/D86030

UltraBlame original commit: e196f9cc2bb23d369daa3f28ab8a73d3634966bb