Bug 1657404 - Refactor gamepad monitoring - Part 1 r=daoshengmu
The starting/stopping of gamepad monitoring is probably a decision that
should be made at the level of the GamepadPlatformService, not in the IPC
actor. This is step 1, later I will expand some of these functions because
it makes the code easier to understand.

Differential Revision: https://phabricator.services.mozilla.com/D86266

UltraBlame original commit: 550630983bc5ad266e8f6ce4e038db8f2bf2c6e6