Bug 1657817 - Fix initGCThing for functions with many TDZ arguments. r=jandem
Check that we will actually need to fill slots before calling
getFixedSlotOffset. This fixes some sanity check asserts, but the issue is
benign in optimized builds because the invalid `offset` is not used if there
are not slots to fill.

Differential Revision: https://phabricator.services.mozilla.com/D86383

UltraBlame original commit: bdfffd302949038754aa1716f93e44c1b7a770d8