Bug 1657677: Resolves FileNotFoundError when running python-test from PyCharm r=firefox-build-system-reviewers,rstewart
When subprocessing to get the venv from pipenv with PyCharm, the returned path has a command sequence to reset the terminal color.
This command sequence is unexpected, and causes the returned path to be incorrect.

The root issue is that colorama mistakenly believes that, if it's ever running underneath PyCharm, it's probably attached to a tty.
However, this assumption isn't true if PyCharm is running a script which subprocesses out a colorama-enabled script.
This issue has already been reported here: https://github.com/tartley/colorama/issues/263

To workaround this issue, we clear the "PYCHARM_HOSTED" environment variable when we invoke pipenv, which works around the colorama logic.

Differential Revision: https://phabricator.services.mozilla.com/D86242

UltraBlame original commit: 467580a245539dd05401e0d1061dac8b6a5dd8bd