Bug 1647899 - Retain or remove search mode across tabs. r=harry
This is very simple, but it gets the job done. It keeps a map in `UrlbarInput`
from browsers to search mode objects. The map is updated in `setSearchMode`, and
it's used to set the search mode in `_afterTabSelectAndFocusChange`.

Differential Revision: https://phabricator.services.mozilla.com/D86296

UltraBlame original commit: 25280cb70b0f4a55c43532b7f1158be47e6b2627