Bug 1657521 - P1. Add method to generate a vpcC box from the VP9 bitstream. r=jolin
The mac VP9 decoder; like the H264 requires some out of band settings before it can be created.

This information is only found in the mp4 container, we can create it from the vp9 bitstream.

For now we ignore the colors information as we can't handle it properly yet in our compositor and this is not available in the bytestream.

Differential Revision: https://phabricator.services.mozilla.com/D86542

UltraBlame original commit: 99693abc449917ff26e9c76eedc100c0ee94793c