Bug 1657374 - Reject v128 if simd disabled. r=rhunt
This adds guards in readValType() and in the validator, and a test
case that checks that the guards are working if SIMD is disabled.

Nearly everything was in place already and this is almost all
plumbing; the exception is the new guard in WasmValidate.h.

Differential Revision: https://phabricator.services.mozilla.com/D86949

UltraBlame original commit: ee8e171958c65e7a0807b608ae4709bed92bde64