Bug 1655406 - Fix race condition when hitting Enter. r=Honza.
Sometimes, it might happen that the user triggered the popup
opening (by typing a letter) and hit Enter before the popup
was already shown, which would result in a blank popup being opened.
This patch addresses that by storing the popup opening promise
and if it exists in the code that handle the Enter key, wait until
the promise resolves and hide the popup.

It would be tricky to write a consistent test as this race is depending
on the server response time and the time it take for the popup to be
shown; but manual tests seemed to indicate this fixes the issue.

Differential Revision: https://phabricator.services.mozilla.com/D86548

UltraBlame original commit: 2c15c7f0d3c6926b77336d07ef49215d9bc01bea