Bug 1386418 - Fix two issues in nsIFrame::HandleRelease. r=dholbert
The frame selection changes can run script, so we need to check for the
frame itself getting destroyed. This fixes the crash as reported.

Additionally, the document check it does for pointer capture is
incorrect, it should use the composed, not uncomposed doc, so that it
works in shadow dom.

Differential Revision: https://phabricator.services.mozilla.com/D87097

UltraBlame original commit: e140c6cd2eb944579ec99458638879c9b8501920