Bug 1659046 - Always add the implementation for SearchFieldCellWithFocusRing. r=mstange
Before bug 1421088, it was only used on x86_64, but now it's always
used, but only implemented on x86_64. It hasn't been a problem so far
because we haven't built on non-x86_64 for a while, but with arm64
coming, things change.

Differential Revision: https://phabricator.services.mozilla.com/D87025

UltraBlame original commit: 5a258bf5a02bfc3795e213df9f9af2f09ddf588e