Bug 1200896 - Force frame flushes when needed to fix browser pioneer test r=rhelmer
When a modal dialog is cancelled, the inertness for other elements
is reverted. However, in order to have the new state (non-inert)
effective, Firefox needs to do a frame flush. This flush is taken
place when it's really needed.

In browser_pioneer_us.js, we have some usage of some buttons when
the flush hasn't taken place yet, so the test fails because the
buttons are not clickable. To fix the test, we add a
getBoundingClientRect() call to force frame flushes to the
corresponding buttons.

Differential Revision: https://phabricator.services.mozilla.com/D86877

UltraBlame original commit: b0add3579bcc35ed2bdf8749f230432acd8092b2