Bug 1659201 create AudioParamMap on demand for AudioWorkletNode#parameters r=pehrsons
When constructing a JS object derived from AudioWorkletNode, the reflector
should be created using the prototype for the derived class.
https://searchfox.org/mozilla-central/rev/2f9eacd9d3d995c937b4251a5557d95d494c9be1/dom/bindings/BindingUtils.h#3078-3081
The reflector is typically created with the correct prototype after
AudioWorkletNode::Constructor() returns.

AudioParamMap_Binding::MaplikeHelpers::Set() calls GetOrCreateDOMReflector()
to get the JS reflector for the AudioParamMap, which obtains the associated
global from the reflector on the AudioWorkletNode parent.
https://searchfox.org/mozilla-central/rev/a315a1a0f09550e23e4590a77e74f36543315da3/dom/bindings/BindingUtils.h#1591-1599
While this was done during the AudioWorkletNode::Constructor(), the
AudioWorkletNode reflector did not yet exist and so was created on-demand
without an explicit prototype, which resulted in the base class prototype
being used incorrectly.
https://searchfox.org/mozilla-central/rev/a315a1a0f09550e23e4590a77e74f36543315da3/dom/bindings/BindingUtils.h#1550

Delaying set-up of the AudioParamMap provides that the AudioWorkletNode
construction is complete and has a correct reflector before AudioParamMap
usage attempts to get the parent reflector.

Differential Revision: https://phabricator.services.mozilla.com/D87554

UltraBlame original commit: ee0b3cf194830ab789becf129c7dfbc135a67ff6