Bug 1659817 - Gracefully handle failure to get AsyncShutdownService in MediaShutdownManger. r=froydnj
Gracefully handle failure to get the shutdown service by returning null when
trying to fetch a barrier rather than asserting. We then gracefully handle the
failure to get a barrier when adding a shutdown blocker. We still assert when
removing a blocker -- my thinking in this case is that we're failing to get the
service because we're already shutdown, however, if we have a blocker added we
should never shutdown until it's removed.

Differential Revision: https://phabricator.services.mozilla.com/D87616

UltraBlame original commit: 69158f3aa05b66da359c6d274733b785c6b912e4