Bug 1661026 [wpt PR 25229] - [scroll-animations] Support animation-timeline:none, a=testonly
Automatic update from web-platform-tests
[scroll-animations] Support animation-timeline:none

Parsing of the 'none' keyword was already supported, so this CL
simply connects the dots between the parsed timeline name
and the timeline we pass to the CSSAnimation.

For the InertEffect, the inherited_time now needs to be a base::nullopt
to prevent an effect value from being produced, so a small adjustment
was needed there.

Note: Changing from IsScrollTimeline() to IsMonotonicallyIncreasing()
is just a cosmetic change.

Note also that the test animation-timeline-in-keyframe.html required
a small adjustment: previously, the computed value of animation-timeline
was 'bar', but there was no scroll-timeline corresponding to 'bar',
so according to spec the animation has no timeline in this case,
and therefore no effect value [1]. We previously fell back
to the DocumentTimeline, which is not correct: we should have fallen
back to "no timeline", which is what we do as of this CL.

[1] https://drafts.csswg.org/css-animations-2/#animation-timeline

Bug: 1074052, 1097046
Change-Id: Ia664f92f70dff68a17eb82315d6bb7bf18de79f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2367835
Reviewed-by: Kevin Ellis <keverschromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruudchromium.org>
Cr-Commit-Position: refs/heads/master{#801363}

--

wpt-commits: 1a64643e528b76e668b8d52f8f35b7a1e68304d6
wpt-pr: 25229

UltraBlame original commit: 9511399fbc489196888287b40ed92db36696b81e