Bug 1659005 part 2: Make nsPageSequenceFrame determine its page range information earlier on, at the start of reflow. r=TYLin
This patch shouldn't change behavior.

After this change, we'll be able to reason about the page range during reflow
(in a later patch in this series). The old place where we determine the
page-range information -- in nsPageSequenceFrame::StartPrint -- unfortunately
runs *after* reflow.  So that was running too late for the information to be
useful when we're laying out pages on sheets.

Differential Revision: https://phabricator.services.mozilla.com/D88469

UltraBlame original commit: fd0c31c45692576b6035f0fdf66ac2bd1e035948