Bug 1645373 - Part 2: Change MInstanceOf to use MDefinition for the prototype. r=jandem
Change `MInstanceOf` to use `MDefinition` for the prototype in preparation for
part 3. Lowering uses `useRegisterOrConstant()` to ensure we still bake in the
prototype for constant objects, which is the default in both Ion and Warp.


Drive-by change:
Remove bogus `LInstanceOfV::lhs()` and `LInstanceOfCache::lhs()` methods. The
left-hand side is a boxed operand, so it can't be accessed through `getOperand`.

Depends on D88573

Differential Revision: https://phabricator.services.mozilla.com/D88574

UltraBlame original commit: fc32f8858aec739f698a9fe514b3cdf050480772