Bug 1661352: Split up GuardFunApplyArray r=jandem
Prior to the NON_PACKED flag, we had to iterate over each element of the array to verify that it was packed. It made sense to fold the argc check in with that code. Now that we are using the flag, I moved the argc check inside the call op to match other argument formats.

Differential Revision: https://phabricator.services.mozilla.com/D88510

UltraBlame original commit: 9abed09112c434ed072ba5f4538e86053f34e97e