Bug 1661495 - [marionette] Don't always set the current content browsing context when registering a new browser. r=marionette-reviewers,maja_zf
Since the patch on bug 1652932 landed in Firefox 80 we always
update the current content browsing context and that now only
when switching to a new window. That leads to an unexpected
change of the current window handle, and as such breaks tests.

Differential Revision: https://phabricator.services.mozilla.com/D88771

UltraBlame original commit: 46f9aa946264098f1af2eae8115954d347d65a8b