Bug 1662278 - Simplify mozperftest metric names. r=tarek
This patch adds two new options to mozperftest. The --simplify-names argument can be provided to enable simplification of metric names and the --simplify-exclude option allows users to select which metrics to skip in the simplification. A bug relating to setting the default of a list option is also fixed here.

Differential Revision: https://phabricator.services.mozilla.com/D88917

UltraBlame original commit: 51251089fe61709267587ad72f459823a4f24869