Bug 1661720 - Move monochrome detection to nsIPrinter. r=bobowen
This moves the special cases to the printing platform code instead
(where it belongs).

The SupportsMonochrome() implementation is the only thing that I found
to work for the "Microsoft Print to PDF" driver, but other similar
drivers like XPS still claim to support monochrome (even though it's not
true). This should avoid the UI getting in a bad state as described in
the bug though, so may be better?

Depends on D89306

Differential Revision: https://phabricator.services.mozilla.com/D89312

UltraBlame original commit: 8cbbcc11ba1bd776c150f69afe5b4d505f4d8534