Bug 1662688 part 5 - Emit GuardFunctionScript for scripted functions. r=iain
Because guarding on the Class and the BaseScript is quite a bit slower than
GuardSpecificFunction, try the function guard if we're attaching the first stub.

Make trial inlining ignore the unused stub so we can still inline. This is similar
to the heuristics WarpOracle uses but is still a bit more conservative.

Differential Revision: https://phabricator.services.mozilla.com/D89140

UltraBlame original commit: a6f52a53ea98e248954bd7504c527eb8cc17a08d