Bug 1663847 part 1 - Add Debugger.Script.parameterNames getter. r=tcampbell
Simplifies the implementation of Debugger.Object.parameterNames (for functions)
and reuses it for scripts. The test is based on the Object-parameterNames.js
test with some minor changes and the documentation is based on the Object docs.

Devtools code will use this in the next patch.

Differential Revision: https://phabricator.services.mozilla.com/D89700

UltraBlame original commit: 530e5373b9acdef4943c157aa580194a6a4fd4cc